pyproject.toml
[project]
name = "sawt"
version = "0.1.0"
description = "Arabic Restaurant Ordering Agent - Multi-agent Saudi Arabic dialect chatbot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "pinecone>=5.0",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pytz>=2024.1",
    "alembic>=1.13",
    "sqlalchemy>=2.0",
    "tiktoken>=0.7",
    "streamlit>=1.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
sawt = "sawt.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sawt"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
strict = true
